feat: Add --fetch
dico is now capable of fetching arbitrary words from any source, called
"provider". Only a French provider is available now, but more will
come.
Words fetched from a provider are stored locally as YAML files in the
tree, which allows for versioning of these files (another goal of this
project).
Once you have all the words you want locally, you can generate a
database from them for (way) faster search.
The database and the dico binary are the only things that should be
packaged and distributed to standard end-users.